tg-6603 [BE] Modificación de verificaciones del perímetro cefálico

+++ back-end/clinic-history/src/main/java/ar/lamansys/sgh/clinichistory/infrastructure/input/rest/ips/constraints/HeadCircumferenceValidator.java
@@ -7,14 +7,14 @@ import javax.validation.ConstraintValidatorContext;
 
 public class HeadCircumferenceValidator implements ConstraintValidator<HeadCircumferenceDataValid, ClinicalObservationDto> {
 
-	private static final Integer MIN_VALUE = 1;
-	private static final Integer MAX_VALUE = 100;
+	private static final Float MIN_VALUE = 1f;
+	private static final Float MAX_VALUE = 100f;
 
 	@Override
 	public boolean isValid(ClinicalObservationDto clinicalObservationDto, ConstraintValidatorContext context){
 		if (clinicalObservationDto != null && clinicalObservationDto.getValue() != null) {
 			try {
-				Integer anthropometricValue = Integer.parseInt(clinicalObservationDto.getValue());
+				Float anthropometricValue = Float.parseFloat(clinicalObservationDto.getValue());
 				return anthropometricValue >= MIN_VALUE && anthropometricValue <= MAX_VALUE;
 			} catch (NumberFormatException exc) {
 				return false;
